update: Organized all of the components into pages for better maintainability
