+++ src/components/home/InstructionsCard.tsx
